apply discount on products and place order with discounted price

+++ order-service/src/main/java/de/tekup/orderservice/dto/OrderResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,5 +14,5 @@ public class OrderResponse {
     private String orderNumber;
     private String orderStatus;
     private List<OrderLineItemsResponse> items;
-    private float totalPrice;
+    private BigDecimal totalPrice;
 }
